refactor: extract partition balance planner
This will let us generate the latest reblancing plan independently of
even a dry-run request, so we can use it in e.g. reporting the live
balanced/unbalanced status for each partition.
